Move SystemTestReporting to reports package
