+++ java/system-test/system-test-suite/src/main/java/sleeper/systemtest/suite/dsl/reports/SystemTestReporting.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package sleeper.systemtest.suite.dsl;
+package sleeper.systemtest.suite.dsl.reports;
 
 import sleeper.systemtest.drivers.compaction.CompactionReportsDriver;
 import sleeper.systemtest.drivers.ingest.IngestReportsDriver;
@@ -22,8 +22,6 @@ import sleeper.systemtest.drivers.instance.ReportingContext;
 import sleeper.systemtest.drivers.instance.SleeperInstanceContext;
 import sleeper.systemtest.drivers.partitioning.PartitionReportDriver;
 import sleeper.systemtest.drivers.util.TestContext;
-import sleeper.systemtest.suite.dsl.reports.SystemTestCompactionJobsReport;
-import sleeper.systemtest.suite.dsl.reports.SystemTestIngestJobsReport;
 import sleeper.systemtest.suite.fixtures.SystemTestClients;
 
 public class SystemTestReporting {
